Cleaning up routes.js
